+++ CHANGELOG.md
@@ -6,7 +6,21 @@ El formato sigue las recomendaciones de Keep a Changelog
 y el versionado utiliza Semantic Versioning.
 
 ---
-## [Sin publicar] - 26/02/2026
+
+## [0.0.2] - 2026-03-27
+
+### Añadido
+- Implementación de la pantalla de login (*Login Screen*).
+- Configuración de la Clean Arquitecture
+- Navegación entre pantallas con Navigation Component
+- Viewmodel del Login
+- Componentes reutilizables (AnimaTextField, RegistrationProgressBar)
+
+### Modificado
+- La estetica de pantalla de bienvenida.
+
+---
+## [0.0.1] - 26/02/2026
 
 ### Añadido
 - Pantalla Welcome